chore: convert more user endpoints to use authentication model
