+++ src/controllers/user/playlists/routes/{id}/cover.js
@@ -1,11 +1,10 @@
 
 const { Playlist } = require('../../../../../db/models')
 const { processFile } = require('../../../../../util/process-file')
+const authenticate = require('../../../authenticate')
 
 module.exports = function () {
-  const operations = {
-    PUT
-  }
+  const operations = { PUT: [authenticate, PUT] }
 
   async function PUT (ctx, next) {
     try {
