+++ src/controllers/user/playlists/routes/{id}/items/index.js
@@ -1,8 +1,9 @@
 const { Playlist, PlaylistItem, Track } = require('../../../../../../db/models')
+const authenticate = require('../../../../authenticate')
 
 module.exports = function () {
   const operations = {
-    PUT,
+    PUT: [authenticate, PUT],
     parameters: [
       {
         name: 'id',
