--- src/controllers/user/stream/api-doc.js
@@ -1,36 +1,0 @@
-const { SwaggerError } = require('../../../util/swagger')
-
-const apiDoc = {
-  swagger: '2.0',
-  info: {
-    title: 'Resonate User Tracks API.',
-    version: '2.0.0-1'
-  },
-  securityDefinitions: {
-    Bearer: { type: 'apiKey', name: 'Authorization', in: 'header' }
-  },
-  definitions: {
-    Error: SwaggerError
-  },
-  responses: {
-    BadRequest: {
-      description: 'Bad request',
-      schema: {
-        $ref: '#/definitions/Error'
-      }
-    },
-    NotFound: {
-      description: 'No tracks were found.',
-      schema: {
-        $ref: '#/definitions/Error'
-      }
-    }
-  },
-  paths: {},
-  security: [
-    { Bearer: [] }
-  ],
-  tags: [{ name: 'stream' }]
-}
-
-module.exports = apiDoc
