+++ src/controllers/user/trackgroups/routes/{id}/index.js
@@ -1,6 +1,7 @@
 const { User, TrackGroup, TrackGroupItem, Track, File } = require('../../../../../db/models')
 const { Op } = require('sequelize')
 const coverSrc = require('../../../../../util/cover-src')
+const authenticate = require('../../../authenticate')
 
 const {
   validateParams
@@ -9,9 +10,9 @@ const {
 
 module.exports = function () {
   const operations = {
-    GET,
-    PUT,
-    DELETE,
+    GET: [authenticate, GET],
+    PUT: [authenticate, PUT],
+    DELETE: [authenticate, DELETE],
     parameters: [
       {
         name: 'id',
