+++ src/controllers/user/trackgroups/routes/{id}/privacy.js
@@ -1,10 +1,11 @@
 // TODO this can be consolidated with ./settings.js
 
 const { TrackGroup } = require('../../../../../db/models')
+const authenticate = require('../../../authenticate')
 
 module.exports = function () {
   const operations = {
-    PUT,
+    PUT: [authenticate, PUT],
     parameters: [
       {
         name: 'id',
