+++ src/controllers/user/tracks/routes/index.js
@@ -1,11 +1,12 @@
 const { Resonate: Sequelize, UserMeta, Track, File } = require('../../../../db/models')
 const { Op } = require('sequelize')
 const ms = require('ms')
+const authenticate = require('../../authenticate')
 
 module.exports = function (trackService) {
   const operations = {
-    GET,
-    POST
+    POST: [authenticate, POST],
+    GET: [authenticate, GET]
   }
 
   async function GET (ctx, next) {
