+++ src/controllers/user/tracks/routes/{id}/file.js
@@ -1,10 +1,11 @@
 
 const { Track } = require('../../../../../db/models')
 const { processFile } = require('../../../../../util/process-file')
+const authenticate = require('../../../authenticate')
 
 module.exports = function () {
   const operations = {
-    PUT
+    PUT: [authenticate, PUT]
   }
 
   async function PUT (ctx, next) {
