+++ src/controllers/user/tracks/routes/{id}/index.js
@@ -2,12 +2,13 @@ const { UserMeta, Track, File } = require('../../../../../db/models')
 const { Op } = require('sequelize')
 const coverSrc = require('../../../../../util/cover-src')
 const { validate } = require('../../../../../schemas/tracks')
+const authenticate = require('../../../authenticate')
 
 module.exports = function () {
   const operations = {
-    GET,
-    DELETE,
-    PUT,
+    DELETE: [authenticate, DELETE],
+    GET: [authenticate, GET],
+    PUT: [authenticate, PUT],
     parameters: [
       {
         name: 'id',
